requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-jazzmin==3.0.0
Pillow==10.2.0
python-slugify==8.0.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.1
nltk==3.8.1
